# docs/swagger.yaml
basePath: /
definitions:
...
  domain.Disable2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  domain.Enable2FAResponse:
    properties:
      qr_code_url:
        type: string
      secret:
        type: string
    type: object
  domain.ErrorResponse:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
...
  domain.Login2FARequest:
    properties:
      code:
        type: string
      pre_auth_token:
        type: string
    required:
    - code
    - pre_auth_token
    type: object
  domain.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  domain.LoginResponse:
    properties:
      expires_in:
        type: integer
      pre_auth_token:
        type: string
      token:
        type: string
      two_factor_required:
        type: boolean
    type: object
...
  domain.RegisterRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
      username:
        type: string
    required:
    - email
    - password
    - username
    type: object
  domain.RegisterResponse:
    properties:
      balances:
        items:
          $ref: '#/definitions/domain.WalletBalanceDTO'
        type: array
      email:
        type: string
      user_id:
        type: string
      username:
        type: string
      wallet_id:
        type: string
    type: object
  domain.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      status:
        type: integer
    type: object
...
  domain.Verify2FARequest:
    properties:
      code:
        type: string
    required:
    - code
    type: object
  domain.WalletBalanceDTO:
    properties:
      asset_symbol:
        type: string
      balance:
        type: number
    type: object
...
host: localhost:8080
info:
  contact:
    email: bintangridwan30@gmail.com
    name: Bintang Ridwan Pribadi
  description: |-
    E-Wallet REST API with user auth, top-up, transfer, and transaction history.
    Built with Go, Gin, GORM, PostgreSQL, JWT, and bcrypt.
  title: Ledger Backend API
  version: "1.0"
paths:
  /auth/2fa/disable:
    post:
      consumes:
      - application/json
      description: Disables 2FA by validating the current TOTP code.
      parameters:
      - description: Disable 2FA payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Disable2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA disabled successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Disable 2FA TOTP
      tags:
      - auth
  /auth/2fa/enable:
    post:
      description: Generates a new TOTP secret key and QR code provisioning URL for
        the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: Secret generated successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Enable2FAResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Generate 2FA TOTP secret key
      tags:
      - auth
  /auth/2fa/login:
    post:
      consumes:
      - application/json
      description: Verifies the TOTP code against the pre-auth token and returns the
        final JWT token.
      parameters:
      - description: Login 2FA challenge payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Login2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "400":
          description: Invalid pre-auth token or code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Invalid TOTP code
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Complete 2FA login challenge
      tags:
      - auth
  /auth/2fa/verify:
    post:
      consumes:
      - application/json
      description: Verifies the first TOTP code to confirm scanning and officially
        enables 2FA for the account.
      parameters:
      - description: Verify 2FA payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 2FA enabled successfully
          schema:
            $ref: '#/definitions/domain.SuccessResponse'
        "400":
          description: Invalid code or signature
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Verify and enable 2FA TOTP
      tags:
      - auth
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token valid for 24 hours.
      parameters:
      - description: Login payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "401":
          description: Invalid email or password
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates a new user account and an associated wallet with balance
        0 in a single atomic transaction.
      parameters:
      - description: Registration payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/domain.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.RegisterResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "409":
          description: Email or username already taken
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/domain.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer {token}" to authenticate.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"